Download and unpack extensions (#34)
* initial commit: downloading and unpacking working but unable to gather pertinent info about the extension

* updated extension discovery to require manifest or package file, also added extension:link command

* correct package json detection

* working structure now. Next step env vars for fetch with basic auth

* add back in fetch auth

* use of arrayHasItems and small todos added

+++ packages/common/src/utils/extract-extension.ts
@@ -1,3 +1,5 @@
+/* eslint-disable */
+
 import _ from 'lodash';
 import fse from 'fs-extra';
 import decompress from 'decompress';
@@ -5,22 +7,31 @@ import path from 'path';
 
 import {FileStructureError} from '../errors';
 import {discoverExtension, IExtensionMeta} from './extension-versions';
+import {arrayHasItems} from './array-has-items';
 
 export const extractExtension = async (archivePath: string, outputDir: string): Promise<IExtensionMeta> => {
     const outputFiles = await decompress(archivePath, outputDir);
 
-    // determine output dir filename from the shortest directory string path
+    let uniqueTopLevelNames: string[];
+    // determine output dir filename from the shortest directory string path if possible
+    // @todo: this wasn't as reliable as I hoped, didn't detect any 'directories' when extracting an example extension. Need a rethink..
     const outputTopLevelDir = _.reduce(
         _.filter(outputFiles, (file) => file.type === 'directory'),
         (a, b) => (a.path.length <= b.path.length ? a : b),
     );
 
     if (!outputTopLevelDir) {
-        await Promise.all(_.map(outputFiles, (file) => fse.remove(path.join(outputDir, file.path))));
-        throw new FileStructureError(`Unexpected file structure after unpacking`);
-    }
+        // @todo: determine output from spitting the filename to the get the top level dir (should be consistent). Not the best but works for now...
+        uniqueTopLevelNames = [...new Set(_.map(outputFiles, (file) => file.path.split(path.sep)[0]))];
 
-    const extractedDistPath = path.join(outputDir, outputTopLevelDir.path);
+        if (!arrayHasItems(uniqueTopLevelNames) || uniqueTopLevelNames.length > 1) {
+            // @todo: this may not remove all the files for the same reason - outputFiles .path property may not contain a reference to directories so only deleting files within. Will be the case in this file and download-neo4j.
+            await Promise.all(_.map(outputFiles, (file) => fse.remove(path.join(outputDir, file.path))));
+            throw new FileStructureError(`Unexpected file structure after unpacking`);
+        }
+    }
+    const outputTopLevel = outputTopLevelDir ? outputTopLevelDir.path : uniqueTopLevelNames![0];
+    const extractedDistPath = path.join(outputDir, outputTopLevel);
 
     // check if this is an extension...
     try {
